+++ bin/compare-sequences.py
@@ -7,6 +7,9 @@ import argparse
 from os.path import join
 from math import log10
 from subprocess import CalledProcessError
+import multiprocessing
+from tempfile import mkdtemp
+from shutil import rmtree
 
 from dark.dna import compareDNAReads, matchToString
 from dark.fasta import FastaReads
@@ -15,29 +18,61 @@ from dark.reads import (Reads, addFASTACommandLineOptions,
 from dark.process import Executor
 from dark.utils import parseRangeString
 
+MAFFT_DEFAULT_ARGS = '--globalpair --maxiterate 1000 --preservecase'
+MAFFT_ALGORITHMS_URL = (
+    'https://mafft.cbrc.jp/alignment/software/algorithms/algorithms.html')
+NEEDLE_DEFAULT_ARGS = 'auto'
 
-def needle(reads, verbose=False):
+
+def mafft(reads, verbose=False, options=None, threads=None):
+    """
+    Run a MAFT alignment and return the two sequences.
+
+    @param reads: An iterable of two reads.
+    @param verbose: If C{True} print progress info to sys.stderr.
+    @param options: A C{str} of options to pass to mafft.
+    @return: A C{Reads} instance with the two aligned sequences.
+    """
+    tempdir = mkdtemp()
+
+    infile = join(tempdir, 'input.fasta')
+    out = join(tempdir, 'result.fasta')
+
+    reads.save(infile)
+
+    if verbose:
+        print('Running mafft.', file=sys.stderr)
+
+    Executor().execute("mafft --thread %d %s '%s' > '%s'" % (
+        threads, options or '', infile, out))
+
+    # Use 'list' in the following to force reading the FASTA from disk.
+    result = Reads(list(FastaReads(out)))
+    rmtree(tempdir)
+
+    return result
+
+
+def needle(reads, verbose=False, options=None):
     """
     Run a Needleman-Wunsch alignment and return the two sequences.
 
     @param reads: An iterable of two reads.
     @param verbose: If C{True} print progress info to sys.stderr.
+    @param options: Additional options to pass to needle.
     @return: A C{Reads} instance with the two aligned sequences.
     """
-    from tempfile import mkdtemp
-    from shutil import rmtree
+    tempdir = mkdtemp()
 
-    dir = mkdtemp()
-
-    file1 = join(dir, 'file1.fasta')
+    file1 = join(tempdir, 'file1.fasta')
     with open(file1, 'w') as fp:
         print(reads[0].toString('fasta'), end='', file=fp)
 
-    file2 = join(dir, 'file2.fasta')
+    file2 = join(tempdir, 'file2.fasta')
     with open(file2, 'w') as fp:
         print(reads[1].toString('fasta'), end='', file=fp)
 
-    out = join(dir, 'result.fasta')
+    out = join(tempdir, 'result.fasta')
 
     def useStderr(e):
         return "Sequences too big. Try 'stretcher'" not in e.stderr
@@ -45,9 +80,10 @@ def needle(reads, verbose=False):
     if verbose:
         print('Running needle.', file=sys.stderr)
     try:
-        Executor().execute("needle -asequence '%s' -bsequence '%s' -auto "
-                           "-outfile '%s' -aformat fasta" % (
-                               file1, file2, out), useStderr=useStderr)
+        Executor().execute(
+            "needle -asequence '%s' -bsequence '%s' %s "
+            "-outfile '%s' -aformat fasta" % (
+                file1, file2, options or '', out), useStderr=useStderr)
     except CalledProcessError as e:
         if useStderr(e):
             raise
@@ -62,7 +98,7 @@ def needle(reads, verbose=False):
 
     # Use 'list' in the following to force reading the FASTA from disk.
     result = Reads(list(FastaReads(out)))
-    rmtree(dir)
+    rmtree(tempdir)
 
     return result
 
@@ -81,7 +117,27 @@ parser.add_argument(
 
 parser.add_argument(
     '--align', default=False, action='store_true',
-    help='If given, use needle to do a global alignment of the two sequences.')
+    help=('If given, use mafft (the default) or needle (according to the '
+          'algorithm selected by --aligner) to align the two sequences.'))
+
+parser.add_argument(
+    '--aligner', default='mafft', choices=('mafft', 'needle'),
+    help='The alignment algorithm to use.')
+
+parser.add_argument(
+    '--alignerOptions',
+    help=('Optional arguments to pass to the alignment algorithm. If the '
+          'aligner is mafft, the default options are "%s". If needle, "%s". '
+          'Do not try to set the number of threads here - use the --threads '
+          'argument instead. If you are using mafft, see %s for some possible '
+          'option combinations.' %
+          (MAFFT_DEFAULT_ARGS, NEEDLE_DEFAULT_ARGS, MAFFT_ALGORITHMS_URL)))
+
+parser.add_argument(
+    '--threads', type=int, default=multiprocessing.cpu_count(),
+    help=('The number of threads to use when running the aligner (if --align '
+          'is used and the alignment algorithm can make use of multiple '
+          'threads (mafft can, needle cannot)).'))
 
 parser.add_argument(
     '--alignmentFile',
@@ -136,13 +192,27 @@ if args.alignmentFile:
 if args.align:
     len1, len2 = map(len, reads)
     if len1 == len2:
-        print('Pre-alignment, sequence lengths were identical: %s' % len1)
+        print('Pre-alignment, sequence lengths were identical: %d' % len1)
     else:
         print('Pre-alignment, sequence lengths: %d, %d (difference %d)' % (
             len1, len2, abs(len1 - len2)))
 
-    # Align.
-    reads = needle(reads, args.verbose)
+    if args.aligner == 'mafft':
+        # Be careful in examining args.alignerOptions because we want the
+        # user to be able to pass an empty string (so check against None
+        # before deciding to use the default.)
+        options = (MAFFT_DEFAULT_ARGS if args.alignerOptions is None
+                   else args.alignerOptions)
+        reads = mafft(reads, args.verbose, options=options,
+                      threads=args.threads)
+    else:
+        assert args.aligner == 'needle'
+        # Be careful in examining args.alignerOptions because we want the
+        # user to be able to pass an empty string (so check against None
+        # before deciding to use the default.)
+        options = (NEEDLE_DEFAULT_ARGS if args.alignerOptions is None
+                   else args.alignerOptions)
+        reads = needle(reads, args.verbose, options=options)
 
     if args.alignmentFile:
         assert reads.save(args.alignmentFile) == 2